fix: update type in Users database to Long (#8)

+++ src/main/java/dev/nmarulo/depensaapp/app/users/UserService.java
@@ -17,7 +17,7 @@ public class UserService extends BasicServiceImp {
     
     private final UserRepository userRepository;
     
-    public FindByIdUserRes findById(Integer id, User userToken) {
+    public FindByIdUserRes findById(Long id, User userToken) {
         var userId = id;
         
         if (!Objects.equals(userToken.getId(), id)) {
